+++ NEWS.md
@@ -1,8 +1,14 @@
-Xtheme IRC Services 8.0.0-dev5 Development Notes  
+Xtheme IRC Services 8.0.0-dev5 (Git) Development Notes  
 ========================================================  
 nickserv  
 --------  
+* Added FDROPMAIL, which allows Services Operators to FDROP accounts by email address.  
 * Crash bug fixed that was introduced by the option to AKILL on BADMAIL matches, with use of XMLRPC interfaces.  
+
+
+## STAY UP TO DATE WITH XthemeOrg NEWS, UPDATES & SECURITY ADVISORIES  
+Subscribe to the XthemeOrg News at:  
+https://xtheme.org/org-tools/newsletter/   
 
 Xtheme IRC Services 8.0.0-dev4 Development Notes  
 ========================================================